requirements.txt
django==2.2.3
django-crispy-forms==1.7.2
pillow==6.1.0
crispy-forms-materialize == 0.2
